requirements.txt
llama-index==0.10.12
llama-index-vector-stores-chroma==0.1.6
llama-index-embeddings-openai==0.1.7
llama-index-llms-openai==0.1.13
chromadb==0.4.22
python-docx==1.1.0
openai==1.12.0
python-dotenv==1.0.0
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
gdown==4.7.1
requests==2.31.0
